+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.30.16] - 2026-07-03
+
 ### Added
 
 - **extension**: new `Extension::public_api_additions(api, cfg, language)`